cnn_VGG16/requirements.txt
streamlit
tensorflow==2.16.1
pillow
numpy
matplotlib
scikit-learn
gdown
opencv-python-headless
